+++ ros/vive_tracker/launch/node.launch
@@ -4,7 +4,7 @@
   <arg name="frame_id" default="vive" />
   <arg name="port" default="/dev/ttyUSB0" />
 
-  <node pkg="vive_tracker" name="vive_tracker" type="vive_tracker.py" ns="vive_tracker" output="screen" >
+  <node pkg="vive_tracker" name="vive_tracker" type="src/vive_tracker.py" ns="vive_tracker" output="screen" >
     <param name="confidence_thresh" value="$(arg confidence_thresh)" />
 	<param name="frame_id" value="$(arg frame_id)" />
     <param name="port" value="$(arg port)" />
